feat: replace webpack build with hugo pipelines
