+++ .gitignore
@@ -4,5 +4,7 @@
 
 node_modules/
 public/
-theme/
-data/manifest.json
+resources/
+
+.vercel
+hugo_stats.json
