+++ assets/scss/_footer.scss
@@ -27,7 +27,7 @@
 
 	.footer-links {
 		font-size: .875rem;
-		color: black;
+		color: $black;
 
 		.links-header {
 			text-transform: uppercase;
@@ -50,7 +50,7 @@
 
 	.copy {
 		font-family: $family-monospace;
-		font-size: 0.875rem;
+		font-size: .875rem;
 
 		.brand {
 			display: inline-block;
