+++ assets/scss/_header.scss
@@ -7,7 +7,7 @@
 	.navbar-item {
 		padding: .875rem 1rem;
 		font-weight: 300;
-		color: black;
+		color: $black;
 	}
 
 	.brand {
