+++ assets/scss/app.scss
@@ -1,0 +1,16 @@
+@charset 'UTF-8';
+
+@import 'variables';
+
+// Dependencies
+@import 'node_modules/bulma/bulma';
+@import 'node_modules/bulma-checkradio/dist/css/bulma-checkradio';
+@import 'node_modules/@fortawesome/fontawesome-free/scss/fontawesome';
+@import 'node_modules/@fortawesome/fontawesome-free/scss/brands';
+@import 'node_modules/@fortawesome/fontawesome-free/scss/solid';
+
+@import 'common';
+@import 'header';
+@import 'footer';
+@import 'apps';
+@import 'form';
