+++ config.yml
@@ -74,3 +74,7 @@ disableKinds:
   - RSS
   - taxonomy
   - taxonomyTerm
+
+build:
+  noJSConfigInAssets: true
+  writeStats: true
