+++ layouts/_default/baseof.html
@@ -25,7 +25,15 @@
 
 	<link rel="stylesheet" href="//fonts.googleapis.com/css?family=Titillium+Web:400,600,700&subset=latin-ext">
 
-	<link rel="stylesheet" href="{{ index $.Site.Data.manifest "app.css" | relURL }}">
+	{{ $style := resources.Get "scss/app.scss" | toCSS (dict "targetPath" "app.css") | resources.PostCSS | minify | fingerprint | resources.PostProcess }}
+	<link rel="stylesheet" href="{{ $style.Permalink }}">
+
+	{{ $countdown := resources.Get "js/countdown.js" | js.Build }}
+	{{ $form := resources.Get "js/form.js" | js.Build }}
+	{{ $navbar := resources.Get "js/navbar.js" | js.Build }}
+	{{ $js := slice $countdown $navbar $form | resources.Concat "app.js" | babel (dict "minified" true "noComments" true) }}
+
+	<script src="{{ $js.Permalink }}" defer></script>
 
 	<link rel="apple-touch-icon" sizes="180x180" href="{{ "images/favicon/apple-touch-icon.png" | absURL }}">
 	<link rel="icon" type="image/png" sizes="32x32" href="{{ "images/favicon/favicon-32x32.png" | absURL }}">
@@ -39,6 +47,5 @@
 	{{ partial "footer.html" . }}
 
 	{{ template "_internal/google_analytics.html" . }}
-	<script src="{{ index $.Site.Data.manifest "app.js" | relURL }}"></script>
 </body>
 </html>
