+++ package.json
@@ -3,49 +3,20 @@
 	"author": "Code for Romania",
 	"license": "MPL-2.0",
 	"private": true,
-	"scripts": {
-		"lint:css": "sass-lint --verbose",
-		"lint:js": "eslint src/js",
-		"dev": "cross-env NODE_ENV=development webpack --progress --hide-modules --watch",
-		"build": "cross-env NODE_ENV=production webpack --no-progress --hide-modules"
-	},
 	"browserslist": [
-		"last 2 versions",
-		"> 2%"
+		"defaults"
 	],
-	"devDependencies": {
-		"@babel/core": "^7.11.6",
-		"@babel/preset-env": "^7.11.5",
-		"autoprefixer": "^9.8.6",
-		"babel-eslint": "^10.1.0",
-		"babel-loader": "^8.1.0",
-		"cross-env": "^7.0.2",
-		"css-loader": "^4.3.0",
-		"eslint": "^7.10.0",
-		"file-loader": "^6.1.0",
-		"friendly-errors-webpack-plugin": "^1.7.0",
-		"glob-all": "^3.2.1",
-		"json-loader": "^0.5.7",
-		"mini-css-extract-plugin": "^0.11.2",
-		"modernizr": "^3.11.3",
-		"modernizr-loader": "^1.0.1",
-		"node-sass": "^4.14.1",
-		"null-loader": "^4.0.0",
-		"optimize-css-assets-webpack-plugin": "^5.0.4",
-		"postcss-import": "^12.0.1",
-		"postcss-loader": "^4.0.2",
-		"purgecss-webpack-plugin": "^3.0.0",
-		"sass-lint": "^1.13.1",
-		"sass-loader": "^10.0.2",
-		"terser-webpack-plugin": "^4.2.2",
-		"webpack": "^4.44.2",
-		"webpack-cli": "^3.3.12",
-		"webpack-manifest-plugin": "^2.2.0"
-	},
 	"dependencies": {
-		"@fortawesome/fontawesome-free": "^5.15.0",
-		"bulma": "^0.9.1",
+		"@babel/cli": "^7.15.7",
+		"@babel/core": "^7.15.8",
+		"@babel/preset-env": "^7.15.8",
+		"@fortawesome/fontawesome-free": "^5.15.4",
+		"@fullhuman/postcss-purgecss": "^4.0.3",
+		"autoprefixer": "^10.3.7",
+		"bulma": "^0.9.3",
 		"bulma-checkradio": "^1.1.1",
+		"postcss-cli": "^9.0.1",
+		"postcss-import": "^14.0.2",
 		"validate.js": "^0.13.1"
 	}
 }
