--- src/js/app.js
@@ -1,11 +1,0 @@
-let countdown = require('./countdown'),
-	tests = require('./feature-tests'),
-	navbar = require('./navbar'),
-	form = require('./form'),
-	partners = require('./partners');
-
-tests();
-countdown();
-navbar();
-form();
-partners();
