--- src/scss/app.scss
@@ -1,16 +1,0 @@
-@charset "UTF-8";
-
-@import "variables";
-
-// Dependencies
-@import "~bulma/bulma";
-@import "~bulma-checkradio/dist/css/bulma-checkradio";
-@import "~@fortawesome/fontawesome-free/scss/fontawesome";
-@import "~@fortawesome/fontawesome-free/scss/brands";
-@import "~@fortawesome/fontawesome-free/scss/solid";
-
-@import "common";
-@import "header";
-@import "footer";
-@import "apps";
-@import "form";
